requirements.txt
google-cloud-aiplatform>=1.51.0
google-generativeai>=0.4.0
langchain>=0.1.0
langchain-google-vertexai>=0.1.0
pydantic>=2.5.2
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.25.0
python-dotenv>=1.0.0
google-adk>=0.0.0rc4 